+++ src/indexes/helpers/between_group_dispercion.rs
@@ -18,7 +18,7 @@ impl BGD {
     pub fn compute(
         &self,
         x: &ArrayView2<f64>,
-        y: &ArrayView2<usize>,
+        y: &ArrayView2<u32>,
         clusters_centroids: &Vec<Array2<f64>>,
     ) -> Result<Vec<Array2<f64>>, CalcError> {
         zip(y.columns(), clusters_centroids)
@@ -28,7 +28,7 @@ impl BGD {
     fn helper(
         &self,
         x: &ArrayView2<f64>,
-        y: &ArrayView1<usize>,
+        y: &ArrayView1<u32>,
         clusters_centroids: &Array2<f64>,
     ) -> Result<Array2<f64>, CalcError> {
         let (n, d) = x.dim();
